requirements.txt
flask==3.0.0
scikit-learn==1.8.0
sentence-transformers==2.7.0
groq==0.9.0
httpx==0.27.0
pydantic==2.6.0
pymongo==4.6.1
cryptography==42.0.0
numpy==1.26.4
pandas==2.2.0
joblib==1.3.2
python-dotenv==1.0.0
pyarrow==15.0.0
xgboost==3.2.0